+++ src/ds3232.rs
@@ -32,7 +32,7 @@ where
     /// [`enable_32khz_output()`](#method.enable_32khz_output).
     ///
     /// Note: This is only available for DS3232 and DS3234 devices.
-    pub fn enable_32khz_output_on_battery(&mut self) -> Result<(), Error<E, ()>> {
+    pub fn enable_32khz_output_on_battery(&mut self) -> Result<(), Error<E>> {
         let status = self.status | BitFlags::BB32KHZ;
         self.write_status_without_clearing_alarm(status)
     }
@@ -43,7 +43,7 @@ where
     /// it enabled. See [`enable_32khz_output()`](#method.enable_32khz_output).
     ///
     /// Note: This is only available for DS3232 and DS3234 devices.
-    pub fn disable_32khz_output_on_battery(&mut self) -> Result<(), Error<E, ()>> {
+    pub fn disable_32khz_output_on_battery(&mut self) -> Result<(), Error<E>> {
         let status = self.status & !BitFlags::BB32KHZ;
         self.write_status_without_clearing_alarm(status)
     }
@@ -55,10 +55,7 @@ where
     /// temperature changes will not be compensated for.
     ///
     /// Note: This is only available for DS3232 and DS3234 devices.
-    pub fn set_temperature_conversion_rate(
-        &mut self,
-        rate: TempConvRate,
-    ) -> Result<(), Error<E, ()>> {
+    pub fn set_temperature_conversion_rate(&mut self, rate: TempConvRate) -> Result<(), Error<E>> {
         let status = match rate {
             TempConvRate::_64s => self.status & !BitFlags::CRATE1 & !BitFlags::CRATE0,
             TempConvRate::_128s => self.status & !BitFlags::CRATE1 | BitFlags::CRATE0,