+++ src/ds3234.rs
@@ -2,21 +2,17 @@
 use crate::interface::{SpiInterface, WriteData};
 use crate::{ic, BitFlags, Ds323x, Error, Register, TempConvRate, CONTROL_POR_VALUE};
 use core::marker::PhantomData;
-use embedded_hal::{digital::OutputPin, spi};
+use embedded_hal::spi;
 
-impl<SPI, CS, CommE, PinE> Ds323x<SpiInterface<SPI, CS>, ic::DS3234>
+impl<SPI, E> Ds323x<SpiInterface<SPI>, ic::DS3234>
 where
-    SPI: spi::SpiDevice<u8, Error = CommE>,
-    CS: OutputPin<Error = PinE>,
+    SPI: spi::SpiDevice<u8, Error = E>,
 {
     /// Create a new instance.
-    pub fn new_ds3234(spi: SPI, chip_select: CS) -> Self {
+    pub fn new_ds3234(spi: SPI) -> Self {
         const STATUS_POR_VALUE: u8 = BitFlags::OSC_STOP | BitFlags::BB32KHZ | BitFlags::EN32KHZ;
         Ds323x {
-            iface: SpiInterface {
-                spi,
-                cs: chip_select,
-            },
+            iface: SpiInterface { spi },
             control: CONTROL_POR_VALUE,
             status: STATUS_POR_VALUE,
             _ic: PhantomData,
@@ -24,8 +20,8 @@ where
     }
 
     /// Destroy driver instance, return SPI bus instance and CS output pin.
-    pub fn destroy_ds3234(self) -> (SPI, CS) {
-        (self.iface.spi, self.iface.cs)
+    pub fn destroy_ds3234(self) -> SPI {
+        self.iface.spi
     }
 
     /// Enable the 32kHz output when battery-powered. (enabled per default)
@@ -34,7 +30,7 @@ where
     /// [`enable_32khz_output()`](#method.enable_32khz_output).
     ///
     /// Note: This is only available for DS3232 and DS3234 devices.
-    pub fn enable_32khz_output_on_battery(&mut self) -> Result<(), Error<CommE, PinE>> {
+    pub fn enable_32khz_output_on_battery(&mut self) -> Result<(), Error<E>> {
         let status = self.status | BitFlags::BB32KHZ;
         self.write_status_without_clearing_alarm(status)
     }
@@ -45,7 +41,7 @@ where
     /// it enabled. See [`enable_32khz_output()`](#method.enable_32khz_output).
     ///
     /// Note: This is only available for DS3232 and DS3234 devices.
-    pub fn disable_32khz_output_on_battery(&mut self) -> Result<(), Error<CommE, PinE>> {
+    pub fn disable_32khz_output_on_battery(&mut self) -> Result<(), Error<E>> {
         let status = self.status & !BitFlags::BB32KHZ;
         self.write_status_without_clearing_alarm(status)
     }
@@ -57,10 +53,7 @@ where
     /// temperature changes will not be compensated for.
     ///
     /// Note: This is only available for DS3232 and DS3234 devices.
-    pub fn set_temperature_conversion_rate(
-        &mut self,
-        rate: TempConvRate,
-    ) -> Result<(), Error<CommE, PinE>> {
+    pub fn set_temperature_conversion_rate(&mut self, rate: TempConvRate) -> Result<(), Error<E>> {
         let status = match rate {
             TempConvRate::_64s => self.status & !BitFlags::CRATE1 & !BitFlags::CRATE0,
             TempConvRate::_128s => self.status & !BitFlags::CRATE1 | BitFlags::CRATE0,
@@ -73,14 +66,14 @@ where
     /// Enable the temperature conversions when battery-powered. (enabled per default)
     ///
     /// Note: This is only available for DS3234 devices.
-    pub fn enable_temperature_conversions_on_battery(&mut self) -> Result<(), Error<CommE, PinE>> {
+    pub fn enable_temperature_conversions_on_battery(&mut self) -> Result<(), Error<E>> {
         self.iface.write_register(Register::TEMP_CONV, 0)
     }
 
     /// Disable the temperature conversions when battery-powered.
     ///
     /// Note: This is only available for DS3234 devices.
-    pub fn disable_temperature_conversions_on_battery(&mut self) -> Result<(), Error<CommE, PinE>> {
+    pub fn disable_temperature_conversions_on_battery(&mut self) -> Result<(), Error<E>> {
         self.iface
             .write_register(Register::TEMP_CONV, BitFlags::TEMP_CONV_BAT)
     }